backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
supabase==2.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
google-api-python-client==2.130.0
openpyxl==3.1.2
pandas==2.2.2
